ovs-base: Create new ovs-base image
This patch creates a new ovs-base image which handles the actual
building of the OVS source code.  This significantly simplifies the
related docker files.  Furthermore, by cleverly organizing the
ovs-base image so that intermediate files are created and deleted in
the same command, this patch significantly reduces the amount of data
transfered to the docker repository.